Alias could not be created.
On a clean environment where the configured alias and index
did not already exist, the alias could not be created
because the method was trying to add an index that did not exist
to the alias.

The change ensures that an index with the correct name exists prior
to the alias being created.