Initial implementation of playlist processing
This currently retrieves the playlists from a
single user whose user id is hardcoded into
the AssetBankHarvester class.

This can be changed at a later date if playlists
should be taken from any user.

Implementations interacting with other systems
should return a list containing an object for
each playlist that a video is part of:

[
  {
    name: <playlist name>,
    position: <playlist position>
  },
  {
    name: <playlist name>,
    position: <playlist position>
  }
]

This will be added as a nested object on the main video
document created in Elasticsearch and allow a client to
recreate playlists with correct ordering.