Add a new adaptor to orchestrate transcript data
In the current implemntation large text-based transcript data
is not working well with Elasticsearch.

New adaptor handles this data differently and stores it in it's own repository on the server where it can be served by the application directly.

As a result adaptors should now specify a list of fields that
they want to retrieve from the harvested data.